django_backend/requirements.txt
# Requirements

django>=2.1.2
djangorestframework==3.8.2
django-cors-headers
channels
